requirements.txt
Flask~=2.3.2
langchain~=0.0.194
python-dotenv~=1.0.0
docs~=0.1.0
openai~=0.27.8
tenacity~=8.2.2
infinopy~=0.0.1
matplotlib~=3.7.1
config~=0.5.1
requests~=2.31.0
streamlit~=1.24.0
numpy~=1.25.0